+++ CHANGELOG.md
@@ -8,6 +8,8 @@ NEW FEATURES
   `ActiveRecord::Type`, not just the constants defined under it
 - If `type` is specified but serialization options aren't then attempt to
   detect a default serializer based on the type.
+- New serializer for `ipaddr`, which acts as a default for `inet` and
+  `cidr` too.
 
 BREAKING CHANGES
 - Actually drop support for rails 4.x, we should have done this in 0.7.0